QA/server/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
pymysql==1.1.1
PyJWT==2.10.1
langchain==0.3.14
langchain-community==0.3.14
langchain-chroma==0.2.2
langchain-ollama==0.3.0
chromadb==0.6.3
python-docx==1.1.2
pypdf==5.1.0
langchain-openai==0.2.14
openai==1.58.1